requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
scipy==1.12.0
six==1.16.0
text-unidecode==1.3
